+++ cffi_module/cffi_defs.py
@@ -236,6 +236,12 @@ extern int msr3_pack (const MS3Record *msr,
                       void *handlerdata, int64_t *packedsamples,
                       uint32_t flags, int8_t verbose);
 
+typedef struct MS3RecordPacker MS3RecordPacker;
+
+extern MS3RecordPacker *msr3_pack_init (const MS3Record *msr, uint32_t flags, int8_t verbose);
+extern int msr3_pack_next (MS3RecordPacker *packer, char **record, int32_t *reclen);
+extern void msr3_pack_free (MS3RecordPacker **packer, int64_t *packedsamples);
+
 extern int msr3_repack_mseed3 (const MS3Record *msr, char *record, uint32_t recbuflen, int8_t verbose);
 
 extern int msr3_repack_mseed2 (const MS3Record *msr, char *record, uint32_t recbuflen, int8_t verbose);
@@ -301,6 +307,15 @@ extern int mstl3_resize_buffers (MS3TraceList *mstl);
 extern int64_t mstl3_pack (MS3TraceList *mstl, void (*record_handler) (char *, int, void *),
                            void *handlerdata, int reclen, int8_t encoding,
                            int64_t *packedsamples, uint32_t flags, int8_t verbose, char *extra);
+
+typedef struct MS3TraceListPacker MS3TraceListPacker;
+
+extern MS3TraceListPacker *mstl3_pack_init (MS3TraceList *mstl, int reclen, int8_t encoding,
+                                            uint32_t flags, int8_t verbose, char *extra,
+                                            uint32_t flush_idle_seconds);
+extern int mstl3_pack_next (MS3TraceListPacker *packer, uint32_t flags, char **record, int32_t *reclen);
+extern void mstl3_pack_free (MS3TraceListPacker **packer, int64_t *packedsamples);
+
 extern int64_t mstl3_pack_ppupdate_flushidle (MS3TraceList *mstl,
                                               void (*record_handler) (char *, int, void *),
                                               void *handlerdata, int reclen, int8_t encoding,