+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `MS3Record.generate()`, a generator to produce miniSEED records
+
+### Deprecated
+- `MS3Record.pack()` functionality, use `MS3Record.generate()` instead.
 
 ### Changed
 - libmseed updated to v3.2.0